second fragment adding coordinate on image done

+++ app/src/main/java/com/example/madcamp/MapCoord.java
@@ -4,7 +4,7 @@ import com.google.android.material.floatingactionbutton.FloatingActionButton;
 
 public class MapCoord {
     boolean valid = false;
-    Float longitude;
-    Float latitude;
+    Double longitude;
+    Double latitude;
 
 }
